serializersProject1/requirements.txt
absl-py==2.1.0
anyio==3.7.1
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
audioread==3.0.1
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==7.1.2
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
Django==4.2
django-extensions==3.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
executing==2.0.1
Faker==25.0.1
ffmpeg-python==0.2.0
flatbuffers==24.3.25
fonttools==4.51.0
future==1.0.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gprof2dot==2022.7.29
grpcio==1.63.0
h11==0.12.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
ipython==8.23.0
jedi==0.19.1
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.38.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
moviepy==1.0.3
msgpack==1.0.8
mutagen==1.47.0
norbert==0.2.1
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.2
openai==1.23.6
opt-einsum==3.3.0
packaging==24.0
pandas==1.5.3
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.1
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==3.20.3
psycopg==3.1.18
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytube==15.0.0
pytz==2024.1
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
resampy==0.4.3
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
spleeter==2.3.2
sqlparse==0.4.4
stack-data==0.6.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.1
termcolor==2.4.0
threadpoolctl==3.5.0
tomli==2.0.1
toposort==1.10
tqdm==4.66.4
traitlets==5.14.2
typer==0.3.2
typing_extensions==4.5.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yt-dlp==2024.4.9
